\encoding{UTF-8}

\name{pop.lion-package}
\alias{pop.lion-package}
\alias{pop.lion}
\docType{package}
\title{
Lion population models
}
\description{
A package to run simulations of lion populations using an Individual-Based Model compiled in C.
}
\details{
\tabular{ll}{
Package: \tab pop.lion\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2020-04-28\cr
License: \tab GPL-3\cr
}
}
\author{
Guillaume Chapron <gchapron@carnivoreconservation.org> with contributions from Matthew Wijers, Andrew Loveridge and David Macdonald.
}
